requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
coverage==5.0.3
cryptography==2.8
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
loguru==0.4.1
MarkupSafe==1.1.1
peewee==3.13.1
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.5
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
win32-setctime==1.0.1
wincertstore==0.2
